added a third page for unsorted. sorted topic_ids.
